adds the missing -linkall option when unix.cmxs is created
When a cmxs file is created from the cmxa library without the -linkall
option the resulting cmxs file will not include all symbols that are
presented in the cmxa file. As a result, dynamic linking wouldn't
always work, a prominent example is the `unix_exit` symbol that is not
present in unix.cmxs (but is presenet in unix.cmxa). This symbol is
never referenced in the compiler/runtime code, but is used in
lwt.unix (directly via the external directive). As a result it is not
possible to dynamically link lwt.unix library.